rename scan to 'triple_at_scan'
