+++ pg_search/src/postgres/customscan/triple_at_scan/mod.rs
@@ -26,8 +26,10 @@ use crate::postgres::customscan::builders::custom_scan::CustomScanBuilder;
 use crate::postgres::customscan::builders::custom_state::{
     CustomScanStateBuilder, CustomScanStateWrapper,
 };
-use crate::postgres::customscan::example::qual_inspect::{can_use_quals, extract_quals, Qual};
 use crate::postgres::customscan::explainer::Explainer;
+use crate::postgres::customscan::triple_at_scan::qual_inspect::{
+    can_use_quals, extract_quals, Qual,
+};
 use crate::postgres::customscan::{node, CustomScan, CustomScanState};
 use crate::postgres::options::SearchIndexCreateOptions;
 use crate::postgres::rel_get_bm25_index;
@@ -41,10 +43,10 @@ use std::ffi::CStr;
 use std::ptr::{addr_of, addr_of_mut};
 
 #[derive(Default)]
-pub struct Example;
+pub struct TripleAtScan;
 
 #[derive(Default)]
-pub struct ExampleScanState {
+pub struct TripleAtScanState {
     snapshot: Option<pg_sys::Snapshot>,
     heaprel: Option<pg_sys::Relation>,
     index_name: String,
@@ -58,9 +60,9 @@ pub struct ExampleScanState {
     score_field_indices: Vec<usize>,
 }
 
-impl CustomScanState for ExampleScanState {}
+impl CustomScanState for TripleAtScanState {}
 
-impl ExampleScanState {
+impl TripleAtScanState {
     #[inline(always)]
     pub fn need_scores(&self) -> bool {
         !self.score_field_indices.is_empty()
@@ -124,9 +126,9 @@ impl PrivateData {
     }
 }
 
-impl CustomScan for Example {
-    const NAME: &'static CStr = c"ParadeDB Example Scan";
-    type State = ExampleScanState;
+impl CustomScan for TripleAtScan {
+    const NAME: &'static CStr = c"ParadeDB Scan";
+    type State = TripleAtScanState;
 
     fn callback(mut builder: CustomPathBuilder) -> Option<pg_sys::CustomPath> {
         if !GUCS.enable_custom_scan() {
@@ -281,7 +283,7 @@ impl CustomScan for Example {
             );
         }
 
-        Example::rescan_custom_scan(state)
+        TripleAtScan::rescan_custom_scan(state)
     }
 
     fn exec_custom_scan(state: &mut CustomScanStateWrapper<Self>) -> *mut pg_sys::TupleTableSlot {
